+++ MauiApp1234/Pages/Quiz/QuizPage6.xaml.cs
@@ -12,7 +12,7 @@ public partial class QuizPage6 : ContentPage
 
     private void Button_Clicked(object sender, EventArgs e)
     {
-        int id = 0;
+        long id = 0;
 
 
 
@@ -27,7 +27,7 @@ public partial class QuizPage6 : ContentPage
             }
             else
             {
-                id = Convert.ToInt16(n);
+                id = Convert.ToInt64(n);
             }
